+++ tapis-appsapi/src/main/java/edu/utexas/tacc/tapis/apps/api/model/JobAttributes.java
@@ -26,6 +26,8 @@ public final class JobAttributes
   public String execSystemExecDir;
   public String execSystemInputDir;
   public String execSystemOutputDir;
+  public String dtnSystemInputDir;
+  public String dtnSystemOutputDir;
   public String execSystemLogicalQueue;
   public String archiveSystemId;
   public String archiveSystemDir;
@@ -61,6 +63,8 @@ public final class JobAttributes
     execSystemExecDir = a.getExecSystemExecDir();
     execSystemInputDir = a.getExecSystemInputDir();
     execSystemOutputDir = a.getExecSystemOutputDir();
+    dtnSystemInputDir = a.getDtnSystemInputDir();
+    dtnSystemOutputDir = a.getDtnSystemOutputDir();
     execSystemLogicalQueue = a.getExecSystemLogicalQueue();
     archiveSystemId = a.getArchiveSystemId();
     archiveSystemDir = a.getArchiveSystemDir();