+++ tapis-appslib/src/test/java/edu/utexas/tacc/tapis/apps/service/AppsServiceTest.java
@@ -297,6 +297,8 @@ public class AppsServiceTest
     app0.setExecSystemExecDir(execSystemExecDir2);
     app0.setExecSystemInputDir(execSystemInputDir2);
     app0.setExecSystemOutputDir(execSystemOutputDir2);
+    app0.setDtnSystemInputDir(dtnSystemInputDir2);
+    app0.setDtnSystemOutputDir(dtnSystemOutputDir2);
     app0.setExecSystemLogicalQueue(execSystemLogicalQueue2);
     app0.setArchiveSystemId(archiveSystemId2);
     app0.setArchiveSystemDir(archiveSystemDir2);
@@ -366,6 +368,8 @@ public class AppsServiceTest
     app0.setExecSystemExecDir(execSystemExecDir2);
     app0.setExecSystemInputDir(execSystemInputDir2);
     app0.setExecSystemOutputDir(execSystemOutputDir2);
+    app0.setDtnSystemInputDir(dtnSystemInputDir2);
+    app0.setDtnSystemOutputDir(dtnSystemOutputDir2);
     app0.setExecSystemLogicalQueue(execSystemLogicalQueue2);
     app0.setArchiveSystemId(archiveSystemId2);
     app0.setArchiveSystemDir(archiveSystemDir2);
@@ -1699,6 +1703,8 @@ public class AppsServiceTest
     Assert.assertEquals(tmpApp.getExecSystemExecDir(), app0.getExecSystemExecDir());
     Assert.assertEquals(tmpApp.getExecSystemInputDir(), app0.getExecSystemInputDir());
     Assert.assertEquals(tmpApp.getExecSystemOutputDir(), app0.getExecSystemOutputDir());
+    Assert.assertEquals(tmpApp.getDtnSystemInputDir(), app0.getDtnSystemInputDir());
+    Assert.assertEquals(tmpApp.getDtnSystemOutputDir(), app0.getDtnSystemOutputDir());
     Assert.assertEquals(tmpApp.getExecSystemLogicalQueue(), app0.getExecSystemLogicalQueue());
     Assert.assertEquals(tmpApp.getArchiveSystemId(), app0.getArchiveSystemId());
     Assert.assertEquals(tmpApp.getArchiveSystemDir(), app0.getArchiveSystemDir());
@@ -1863,6 +1869,8 @@ public class AppsServiceTest
 //    Assert.assertEquals(tmpApp.getExecSystemExecDir(), app0.getExecSystemExecDir());
 //    Assert.assertEquals(tmpApp.getExecSystemInputDir(), app0.getExecSystemInputDir());
 //    Assert.assertEquals(tmpApp.getExecSystemOutputDir(), app0.getExecSystemOutputDir());
+//    Assert.assertEquals(tmpApp.getDtnSystemInputDir(), app0.getDtnSystemInputDir());
+//    Assert.assertEquals(tmpApp.getDtnSystemOutputDir(), app0.getDtnSystemOutputDir());
 //    Assert.assertEquals(tmpApp.getExecSystemLogicalQueue(), app0.getExecSystemLogicalQueue());
 //    Assert.assertEquals(tmpApp.getArchiveSystemId(), app0.getArchiveSystemId());
 //    Assert.assertEquals(tmpApp.getArchiveSystemDir(), app0.getArchiveSystemDir());