+++ CHANGELOG.md
@@ -9,6 +9,18 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
+## 1.6.1 - 2024-??-??
+
+Incremental improvements.
+
+### New features:
+- Add DTN attributes dtnSystemInputDir, dtnSystemInputDir.
+- Use shared code TapisUtils.weaklyValidateUri() for additional validation of file inputs.
+
+### Bug fixes:
+- None.???
+
+---------------------------------------------------------------------------
 ## 1.6.0 - 2024-01-17
 
 Incremental improvements.